{"schema":"decodex/commit/1","summary":"Normalize live baseline checks modules","authority":"manual"}
